# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sila (صِلة) API",
    "description": "Gold marketplace for the Iraqi market. All errors use `{error_code, message, status}`. Money and weights are decimal strings.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ownership/me": {
      "get": {
        "tags": [
          "Ownership"
        ],
        "summary": "Verified gold balance (signature checked before returning), by karat",
        "operationId": "my_ownership_api_ownership_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipOut"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/ownership/resale": {
      "post": {
        "tags": [
          "Ownership"
        ],
        "summary": "Offer part of my holdings on the market (Workflow 09; price set by the server)",
        "operationId": "create_resale_api_ownership_resale_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional. Retrying with the same key returns the original result.",
              "title": "Idempotency-Key"
            },
            "description": "Optional. Retrying with the same key returns the original result."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResaleIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "200": {
            "description": "Idempotent replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Ownership"
        ],
        "summary": "My resale listings (open first)",
        "operationId": "my_resales_api_ownership_resale_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingOut"
                  },
                  "title": "Response My Resales Api Ownership Resale Get"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ownership/resale/{listing_id}": {
      "patch": {
        "tags": [
          "Ownership"
        ],
        "summary": "Suspend, re-activate or withdraw (final, releases unsold grams) my resale listing",
        "operationId": "update_resale_api_ownership_resale__listing_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Listing Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResaleStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Latest in-app notifications (newest first) and the unread count",
        "operationId": "list_notifications_api_notifications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsOut"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark notifications as read (all of them when ids is empty)",
        "operationId": "mark_read_api_notifications_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsOut"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/interest": {
      "post": {
        "tags": [
          "Coming soon"
        ],
        "summary": "Join the waitlist for real estate or oil (public)",
        "operationId": "register_interest_api_interest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__modules__interest__router__InterestOut"
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/admin/interest": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "Waitlist signups",
        "operationId": "interest_api_admin_interest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/app__modules__admin__service__InterestOut"
                  },
                  "type": "array",
                  "title": "Response Interest Api Admin Interest Get"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message",
          "status"
        ],
        "title": "ErrorResponse",
        "description": "Body of every non-2xx response."
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterestIn": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "asset_class": {
            "type": "string",
            "enum": [
              "real_estate",
              "oil"
            ],
            "title": "Asset Class"
          }
        },
        "type": "object",
        "required": [
          "email",
          "asset_class"
        ],
        "title": "InterestIn"
      },
      "KaratBalance": {
        "properties": {
          "karat": {
            "type": "integer",
            "title": "Karat"
          },
          "owned_grams": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Owned Grams"
          },
          "reserved_grams": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Reserved Grams",
            "description": "Offered in open resale listings"
          },
          "available_to_resell_grams": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Available To Resell Grams"
          }
        },
        "type": "object",
        "required": [
          "karat",
          "owned_grams",
          "reserved_grams",
          "available_to_resell_grams"
        ],
        "title": "KaratBalance"
      },
...
      "ListingOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "seller_id": {
            "type": "string",
            "format": "uuid",
            "title": "Seller Id"
          },
          "seller_name": {
            "type": "string",
            "title": "Seller Name",
            "description": "The seller's name; for an investor resale a neutral label, never the name"
          },
          "listing_type": {
            "$ref": "#/components/schemas/ListingType",
            "description": "seller_listing, or investor_resale (show 'إعادة بيع من مستثمر')"
          },
          "seller_kyc_verified": {
            "type": "boolean",
            "title": "Seller Kyc Verified"
          },
          "karat": {
            "type": "integer",
            "title": "Karat"
          },
          "total_weight_grams": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Weight Grams"
          },
          "available_weight_grams": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Available Weight Grams"
          },
          "base_price_per_gram": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Base Price Per Gram",
            "description": "Reference price fixed at creation"
          },
          "current_price_per_gram": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price Per Gram",
            "description": "Live price for this karat now; the purchase uses this (see preview)"
          },
          "status": {
            "$ref": "#/components/schemas/ListingStatus"
          },
          "is_promoted": {
            "type": "boolean",
            "title": "Is Promoted",
            "description": "Effective: is_promoted AND promotion_expiry_date > now"
          },
          "promotion_expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promotion Expiry Date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "seller_id",
          "seller_name",
          "listing_type",
          "seller_kyc_verified",
          "karat",
          "total_weight_grams",
          "available_weight_grams",
          "base_price_per_gram",
          "current_price_per_gram",
          "status",
          "is_promoted",
          "promotion_expiry_date",
          "created_at",
          "updated_at"
        ],
        "title": "ListingOut"
      },
      "ListingStatus": {
        "type": "string",
        "enum": [
          "active",
          "sold_out",
          "suspended",
          "withdrawn"
        ],
        "title": "ListingStatus"
      },
...
      "ListingType": {
        "type": "string",
        "enum": [
          "seller_listing",
          "investor_resale"
        ],
        "title": "ListingType"
      },
...
      "MarkReadIn": {
        "properties": {
          "ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          }
        },
        "type": "object",
        "title": "MarkReadIn"
      },
...
      "NotificationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          },
          "read": {
            "type": "boolean",
            "title": "Read"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "link",
          "read",
          "created_at"
        ],
        "title": "NotificationOut"
      },
      "NotificationsOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "type": "array",
            "title": "Items"
          },
          "unread_count": {
            "type": "integer",
            "title": "Unread Count"
          }
        },
        "type": "object",
        "required": [
          "items",
          "unread_count"
        ],
        "title": "NotificationsOut"
      },
...
      "OwnershipOut": {
        "properties": {
          "investor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Investor Id"
          },
          "total_accumulated_grams": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Accumulated Grams"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "description": "Signature checked on this request"
          },
          "digital_signature_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digital Signature Token"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "by_karat": {
            "items": {
              "$ref": "#/components/schemas/KaratBalance"
            },
            "type": "array",
            "title": "By Karat",
            "description": "Derived from the transaction log; sums to the total"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Shown when the investor owns nothing"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "التوقيع الرقمي إثبات تقني داخلي لسلامة رصيدك داخل منصة صِلة، وليس سند ملكية قانونياً معترفاً به رسمياً."
          }
        },
        "type": "object",
        "required": [
          "investor_id",
          "total_accumulated_grams",
          "verified",
          "digital_signature_token",
          "updated_at"
        ],
        "title": "OwnershipOut"
      },
...
      "ResaleIn": {
        "properties": {
          "karat": {
            "type": "integer",
            "title": "Karat",
            "examples": [
              21
            ]
          },
          "weight_grams": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,7}|(?=[\\d.]{1,11}0*$)\\d{0,7}\\.\\d{0,3}0*$)"
              }
            ],
            "title": "Weight Grams",
            "examples": [
              "84.250"
            ]
          }
        },
        "type": "object",
        "required": [
          "karat",
          "weight_grams"
        ],
        "title": "ResaleIn"
      },
      "ResaleStatusIn": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(active|suspended|withdrawn)$",
            "title": "Status",
            "description": "suspended keeps the grams reserved; withdrawn is final and releases them"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ResaleStatusIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__modules__admin__service__InterestOut": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "asset_class": {
            "type": "string",
            "title": "Asset Class"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "asset_class",
          "created_at"
        ],
        "title": "InterestOut"
      },
      "app__modules__interest__router__InterestOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Same reply for a new or a repeated signup"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "InterestOut"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /api/auth/login",
        "scheme": "bearer"
      }
    }
  },
...
}